OneDrive/Documents/ai-project/terraform-ai-analyzer/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
python-multipart>=0.0.9
python-hcl2>=4.3.0
networkx>=3.3
httpx>=0.27.0
